+++ features/support/app_pages/iva/student_doc.rb
@@ -11,11 +11,11 @@ class StudentDoc < IDmeBase
 
       if populate
 
-        unique_data = data_for(:experian_user) #info for unique user
+        unique_data = data_for(:experian_user) #info for unique and duplicate users
         denied_data = data_for(:fail_experian) #info for denied user
 
         case type
-        when "unique"
+        when "unique", "duplicate"
           populate_fields(data: unique_data)
         when "denied"
           populate_fields(data: denied_data)