+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. This projec
 ## 2.2.6 (07/11/2016)
 - HOTFIX: Update Carhartt partner test
 - HOTFIX: Update Columbia sport test and fix verify_discount helper method
+- QA-250: Dupe attempt test for fireman doc upload
+- QA-252: Dupe attempt test for police officer
+- QA-254: Dupe attempt test for student doc upload
+- HOTFIX: Typo in emt doc upload feature
 
 ## 2.2.5 (07/10/2016)
 - HOTFIX: Update Underarmour partner test